Update to canvas and map class
when new track is added, AllConnected variable set to false again.
Also cleaned up files

+++ canvas.cpp
@@ -1523,11 +1523,20 @@ void Canvas::mousePressEvent(QMouseEvent* event)
             }
         };
         update();
+        if (addedTrack)
+        {
+
+        }
     }
     else if (event->button() == Qt::RightButton)
     {
-        drawnLayout->deleteElement(finalX, finalY);
-        update();
+        if (drawnLayout->checkElementExists(finalX, finalY))
+        {
+            drawnLayout->deleteElement(finalX, finalY);
+            update();
+            drawnLayout->setAllConnected(false);
+        }
+
     }
 }
